pyproject.toml
[project]
name = "telegram-notifier-fastapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14.0"
dependencies = [
    "fastapi>=0.120.1",
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=22.5",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.2",
]
